+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -13,6 +13,8 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', category::class);
+
         $categories = Category::paginate(10);
         return view('admin.categories.index', compact('categories'));
     }
@@ -22,6 +24,8 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', category::class);
+
         return view('admin.categories.create');
     }
 
@@ -30,6 +34,8 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', category::class);
+
         $request->validate([
             'name' => 'required|string|max:255|min:3',
         ]);
@@ -42,6 +48,8 @@ class CategoryController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('view', category::class);
+
         $category = Category::findOrFail($id);
         return view('admin.categories.show', compact('category'));
     }
@@ -51,6 +59,8 @@ class CategoryController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('update', category::class);
+
         $category = Category::findOrFail($id);
         return view('admin.categories.edit',compact('category'));
     }
@@ -60,6 +70,8 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        $this->authorize('update', category::class);
+
         $request->validate([
             'name' => 'required|string|max:255|min:3',
         ]);
@@ -72,6 +84,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize('delete', category::class);
+
         $category->delete();
         return to_route('categories.index')->with('Danger', 'Category deleted successfully');
     }
